requirements.txt
scipy
trl
transformers
accelerate
peft
bitsandbytes
einops
tiktoken
streamlit
pdfminer.six
torch
